pyproject.toml
[project]
name = "fastapi-dope-2026-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.13",
    "botocore>=1.42.13",
    "fastapi[standard]>=0.125.0",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "tortoise>=0.1.1",
]
